extract out the last two components
