+++ app/src/models/diff/index.ts
@@ -1,4 +1,6 @@
+export * from './diff-data'
 export * from './diff-line'
-export * from './image'
+export * from './diff-selection'
 export * from './file-summary'
+export * from './image'
 export * from './raw-diff'
